Move worker mock to `test-utils` to reduce duplication
