+++ packages/cli/src/project-references.test.ts
@@ -1,18 +1,19 @@
-import { createDeferredPromise } from '@metamask/utils';
-import { getFixture, getRelativePath, delay } from '@ts-bridge/test-utils';
+import {
+  getFixture,
+  getMockWorker,
+  getRelativePath,
+} from '@ts-bridge/test-utils';
 import assert from 'assert';
 import chalk from 'chalk';
-import { dirname, join } from 'path';
+import { dirname } from 'path';
 import type { Program, ResolvedProjectReference } from 'typescript';
 import { factory, ScriptTarget, sys } from 'typescript';
 import { fileURLToPath } from 'url';
 import { beforeAll, describe, expect, it, vi } from 'vitest';
-import type { WorkerOptions } from 'worker_threads';
 
 import { getTypeScriptConfig } from './config.js';
 import type { DependencyGraph } from './project-references.js';
 import {
-  parallelise,
   createGraph,
   createProjectReferencesCompilerHost,
   getResolvedProjectReferences,
@@ -25,56 +26,15 @@ beforeAll(() => {
   chalk.level = 0;
 });
 
+// This mock is necessary for the `createProjectReferencesCompilerHost` tests
+// to work.
 vi.mock('worker_threads', async (importOriginal) => {
   // eslint-disable-next-line @typescript-eslint/consistent-type-imports
   const original = await importOriginal<typeof import('worker_threads')>();
 
-  /**
-   * The worker code that runs TypeScript code. This is executed in the worker by
-   * using it as a data URL.
-   */
-  const WORKER_CODE = `
-    import { createRequire } from 'module';
-    import { workerData } from 'worker_threads';
-
-    const filename = '${import.meta.url}';
-    const tsconfig = '${join(
-      dirname(fileURLToPath(import.meta.url)),
-      '..',
-      'tsconfig.json',
-    )}';
-
-    const require = createRequire(filename);
-    const { tsImport } = require('tsx/esm/api');
-
-    tsImport(workerData.fileName, {
-      parentURL: import.meta.url,
-      tsconfig,
-    });
-  `;
-
-  /**
-   * A worker that runs TypeScript code.
-   */
-  class TypeScriptWorker extends original.Worker {
-    /**
-     * Creates a new TypeScript worker.
-     *
-     * @param fileName - The file name of worker to run.
-     * @param options - The worker options.
-     * @returns The TypeScript worker.
-     */
-    constructor(fileName: string | URL, options: WorkerOptions = {}) {
-      options.workerData ??= {};
-      options.workerData.fileName = fileName.toString().replace('.js', '.ts');
-
-      super(new URL(`data:text/javascript,${WORKER_CODE}`), options);
-    }
-  }
-
   return {
     ...original,
-    Worker: TypeScriptWorker,
+    Worker: getMockWorker(original.Worker),
   };
 });
 
@@ -191,212 +151,6 @@ describe('topologicalSort', () => {
       ['a', 'b', 'a'],
       ['c', 'd', 'e', 'c'],
     ]);
-  });
-});
-
-describe('parallelise', () => {
-  it('runs a function in parallel for a dependency graph', async () => {
-    const graph: DependencyGraph<string> = new Map([
-      ['a', ['b', 'c']],
-      ['b', ['d', 'e']],
-      ['c', ['e']],
-      ['d', ['e']],
-      ['e', []],
-    ]);
-
-    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
-    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
-    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
-    const { promise: promiseD, resolve: resolveD } = createDeferredPromise();
-    const { promise: promiseE, resolve: resolveE } = createDeferredPromise();
-
-    const set = new Set<string>();
-
-    const sorted = topologicalSort(graph).stack;
-    const parallelPromise = parallelise(sorted, graph, async (key) => {
-      set.add(key);
-      switch (key) {
-        case 'a':
-          await promiseA;
-          break;
-        case 'b':
-          await promiseB;
-          break;
-        case 'c':
-          await promiseC;
-          break;
-        case 'd':
-          await promiseD;
-          break;
-        case 'e':
-          await promiseE;
-          break;
-        default:
-          throw new Error(`Unknown key: "${key}".`);
-      }
-    });
-
-    // `e` is the only node without dependencies, so it should be resolve
-    // first.
-    expect(set.has('e')).toBe(true);
-    expect(set.has('d')).not.toBe(true);
-    expect(set.has('c')).not.toBe(true);
-
-    resolveE();
-    await delay(1);
-
-    // `d` and `c` are the next nodes in the graph, so they should be resolved
-    // next.
-    expect(set.has('d')).toBe(true);
-    expect(set.has('c')).toBe(true);
-
-    resolveD();
-    await delay(1);
-
-    // `b` depends on `d` and `e`, so it should be resolved once they are.
-    expect(set.has('b')).toBe(true);
-    expect(set.has('a')).not.toBe(true);
-
-    resolveC();
-    resolveB();
-    await delay(1);
-
-    // `a` depends on `b` and `c`, so it should be resolved once they are.
-    expect(set.has('a')).toBe(true);
-
-    resolveA();
-    await parallelPromise;
-  });
-
-  it('limits the tasks to the concurrency limit', async () => {
-    const graph: DependencyGraph<string> = new Map([
-      ['a', []],
-      ['b', []],
-      ['c', []],
-    ]);
-
-    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
-    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
-    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
-
-    const set = new Set<string>();
-
-    const sorted = topologicalSort(graph).stack;
-    const parallelPromise = parallelise(
-      sorted,
-      graph,
-      async (key) => {
-        set.add(key);
-        switch (key) {
-          case 'a':
-            await promiseA;
-            break;
-          case 'b':
-            await promiseB;
-            break;
-          case 'c':
-            await promiseC;
-            break;
-          default:
-            throw new Error(`Unknown key: "${key}".`);
-        }
-      },
-      1,
-    );
-
-    // `a` is the first node in the graph, so it should be resolved first.
-    expect(set.has('a')).toBe(true);
-    expect(set.has('b')).not.toBe(true);
-    expect(set.has('c')).not.toBe(true);
-
-    resolveA();
-    await delay(1);
-
-    // `b` is the next node in the graph, so it should be resolved next.
-    expect(set.has('b')).toBe(true);
-    expect(set.has('c')).not.toBe(true);
-
-    resolveB();
-    await delay(1);
-
-    // `c` is the last node in the graph, so it should be resolved last.
-    expect(set.has('c')).toBe(true);
-
-    resolveC();
-
-    await parallelPromise;
-  });
-
-  it('assumes no dependencies if a value is not in the graph', async () => {
-    const graph: DependencyGraph<string> = new Map([
-      ['a', ['b', 'c']],
-      ['b', ['d', 'e']],
-      ['c', ['e']],
-      ['d', ['e']],
-    ]);
-
-    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
-    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
-    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
-    const { promise: promiseD, resolve: resolveD } = createDeferredPromise();
-    const { promise: promiseE, resolve: resolveE } = createDeferredPromise();
-
-    const set = new Set<string>();
-
-    const sorted = topologicalSort(graph).stack;
-    const parallelPromise = parallelise(sorted, graph, async (key) => {
-      set.add(key);
-      switch (key) {
-        case 'a':
-          await promiseA;
-          break;
-        case 'b':
-          await promiseB;
-          break;
-        case 'c':
-          await promiseC;
-          break;
-        case 'd':
-          await promiseD;
-          break;
-        case 'e':
-          await promiseE;
-          break;
-        default:
-          throw new Error(`Unknown key: "${key}".`);
-      }
-    });
-
-    // `e` is the only node without dependencies, so it should be resolve
-    // first.
-    expect(set.has('e')).toBe(true);
-    expect(set.has('d')).not.toBe(true);
-    expect(set.has('c')).not.toBe(true);
-
-    resolveE();
-    await delay(1);
-
-    // `d` and `c` are the next nodes in the graph, so they should be resolved
-    // next.
-    expect(set.has('d')).toBe(true);
-    expect(set.has('c')).toBe(true);
-
-    resolveD();
-    await delay(1);
-
-    // `b` depends on `d` and `e`, so it should be resolved once they are.
-    expect(set.has('b')).toBe(true);
-    expect(set.has('a')).not.toBe(true);
-
-    resolveC();
-    resolveB();
-    await delay(1);
-
-    // `a` depends on `b` and `c`, so it should be resolved once they are.
-    expect(set.has('a')).toBe(true);
-
-    resolveA();
-    await parallelPromise;
   });
 });
