+++ packages/cli/src/project-references.ts
@@ -1,5 +1,4 @@
 import assert from 'assert';
-import { availableParallelism } from 'os';
 import { relative } from 'path';
 import type {
   CompilerHost,
@@ -114,78 +113,6 @@ export function topologicalSort<Value>(
     stack,
     cycles,
   };
-}
-
-/**
- * Run a function in parallel for each value in the graph. This means that the
- * function is run for each value in the graph, but the order in which the
- * values are processed is determined by the topological sort of the graph.
- * Values that are not dependent on each other, or that are dependent on each
- * other, but the dependencies have been resolved, are processed in parallel.
- *
- * @param sortedValues - The sorted values from the graph.
- * @param graph - The dependency graph.
- * @param fn - The function to run for each value.
- * @param maxConcurrency - The maximum number of concurrent functions to run.
- */
-export async function parallelise<Value>(
-  sortedValues: Value[],
-  graph: DependencyGraph<Value>,
-  fn: (value: Value) => Promise<void>,
-  maxConcurrency = availableParallelism(),
-): Promise<void> {
-  const resolved = new Set<Value>();
-  const queue = [...sortedValues];
-  const running = new Map<Value, Promise<void>>();
-
-  /**
-   * Check if a value is ready to be processed. A value is ready if all its
-   * dependencies have been resolved.
-   *
-   * @param value - The value to check.
-   * @returns Whether the value is ready to be processed.
-   */
-  function isReady(value: Value): boolean {
-    const dependencies = graph.get(value) ?? [];
-    return dependencies.every((dep) => resolved.has(dep));
-  }
-
-  /**
-   * Run a task. This will call the provided function and add the value to the
-   * resolved set when the function has completed.
-   *
-   * @param value - The value to run the function for.
-   * @returns A promise that resolves when the function has completed.
-   */
-  async function run(value: Value): Promise<void> {
-    try {
-      await fn(value);
-      resolved.add(value);
-    } finally {
-      running.delete(value);
-    }
-  }
-
-  while (queue.length > 0 || running.size > 0) {
-    while (running.size < maxConcurrency && queue.length > 0) {
-      const nextTaskIndex = queue.findIndex(isReady);
-      if (nextTaskIndex === -1) {
-        break;
-      }
-
-      const nextTask = queue.splice(nextTaskIndex, 1)[0];
-      assert(nextTask);
-
-      const taskPromise = run(nextTask);
-      running.set(nextTask, taskPromise);
-    }
-
-    // Wait for any running task to complete.
-    /* istanbul ignore else -- @preserve */
-    if (running.size > 0) {
-      await Promise.race(running.values());
-    }
-  }
 }
 
 /**
