+++ packages/cli/src/utils.test.ts
@@ -1,6 +1,15 @@
+import { createDeferredPromise } from '@metamask/utils';
+import { delay } from '@ts-bridge/test-utils';
 import { describe, expect, it } from 'vitest';
 
-import { isObject, getIdentifierName, getDefinedArray } from './utils.js';
+import type { DependencyGraph } from './project-references.js';
+import { topologicalSort } from './project-references.js';
+import {
+  isObject,
+  getIdentifierName,
+  getDefinedArray,
+  parallelise,
+} from './utils.js';
 
 describe('isObject', () => {
   it('returns `true` if the value is an object', () => {
@@ -68,5 +77,211 @@ describe('getDefinedArray', () => {
 
   it('returns an empty array if the array is undefined', () => {
     expect(getDefinedArray(undefined)).toStrictEqual([]);
+  });
+});
+
+describe('parallelise', () => {
+  it('runs a function in parallel for a dependency graph', async () => {
+    const graph: DependencyGraph<string> = new Map([
+      ['a', ['b', 'c']],
+      ['b', ['d', 'e']],
+      ['c', ['e']],
+      ['d', ['e']],
+      ['e', []],
+    ]);
+
+    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
+    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
+    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
+    const { promise: promiseD, resolve: resolveD } = createDeferredPromise();
+    const { promise: promiseE, resolve: resolveE } = createDeferredPromise();
+
+    const set = new Set<string>();
+
+    const sorted = topologicalSort(graph).stack;
+    const parallelPromise = parallelise(sorted, graph, async (key) => {
+      set.add(key);
+      switch (key) {
+        case 'a':
+          await promiseA;
+          break;
+        case 'b':
+          await promiseB;
+          break;
+        case 'c':
+          await promiseC;
+          break;
+        case 'd':
+          await promiseD;
+          break;
+        case 'e':
+          await promiseE;
+          break;
+        default:
+          throw new Error(`Unknown key: "${key}".`);
+      }
+    });
+
+    // `e` is the only node without dependencies, so it should be resolve
+    // first.
+    expect(set.has('e')).toBe(true);
+    expect(set.has('d')).not.toBe(true);
+    expect(set.has('c')).not.toBe(true);
+
+    resolveE();
+    await delay(1);
+
+    // `d` and `c` are the next nodes in the graph, so they should be resolved
+    // next.
+    expect(set.has('d')).toBe(true);
+    expect(set.has('c')).toBe(true);
+
+    resolveD();
+    await delay(1);
+
+    // `b` depends on `d` and `e`, so it should be resolved once they are.
+    expect(set.has('b')).toBe(true);
+    expect(set.has('a')).not.toBe(true);
+
+    resolveC();
+    resolveB();
+    await delay(1);
+
+    // `a` depends on `b` and `c`, so it should be resolved once they are.
+    expect(set.has('a')).toBe(true);
+
+    resolveA();
+    await parallelPromise;
+  });
+
+  it('limits the tasks to the concurrency limit', async () => {
+    const graph: DependencyGraph<string> = new Map([
+      ['a', []],
+      ['b', []],
+      ['c', []],
+    ]);
+
+    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
+    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
+    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
+
+    const set = new Set<string>();
+
+    const sorted = topologicalSort(graph).stack;
+    const parallelPromise = parallelise(
+      sorted,
+      graph,
+      async (key) => {
+        set.add(key);
+        switch (key) {
+          case 'a':
+            await promiseA;
+            break;
+          case 'b':
+            await promiseB;
+            break;
+          case 'c':
+            await promiseC;
+            break;
+          default:
+            throw new Error(`Unknown key: "${key}".`);
+        }
+      },
+      1,
+    );
+
+    // `a` is the first node in the graph, so it should be resolved first.
+    expect(set.has('a')).toBe(true);
+    expect(set.has('b')).not.toBe(true);
+    expect(set.has('c')).not.toBe(true);
+
+    resolveA();
+    await delay(1);
+
+    // `b` is the next node in the graph, so it should be resolved next.
+    expect(set.has('b')).toBe(true);
+    expect(set.has('c')).not.toBe(true);
+
+    resolveB();
+    await delay(1);
+
+    // `c` is the last node in the graph, so it should be resolved last.
+    expect(set.has('c')).toBe(true);
+
+    resolveC();
+
+    await parallelPromise;
+  });
+
+  it('assumes no dependencies if a value is not in the graph', async () => {
+    const graph: DependencyGraph<string> = new Map([
+      ['a', ['b', 'c']],
+      ['b', ['d', 'e']],
+      ['c', ['e']],
+      ['d', ['e']],
+    ]);
+
+    const { promise: promiseA, resolve: resolveA } = createDeferredPromise();
+    const { promise: promiseB, resolve: resolveB } = createDeferredPromise();
+    const { promise: promiseC, resolve: resolveC } = createDeferredPromise();
+    const { promise: promiseD, resolve: resolveD } = createDeferredPromise();
+    const { promise: promiseE, resolve: resolveE } = createDeferredPromise();
+
+    const set = new Set<string>();
+
+    const sorted = topologicalSort(graph).stack;
+    const parallelPromise = parallelise(sorted, graph, async (key) => {
+      set.add(key);
+      switch (key) {
+        case 'a':
+          await promiseA;
+          break;
+        case 'b':
+          await promiseB;
+          break;
+        case 'c':
+          await promiseC;
+          break;
+        case 'd':
+          await promiseD;
+          break;
+        case 'e':
+          await promiseE;
+          break;
+        default:
+          throw new Error(`Unknown key: "${key}".`);
+      }
+    });
+
+    // `e` is the only node without dependencies, so it should be resolve
+    // first.
+    expect(set.has('e')).toBe(true);
+    expect(set.has('d')).not.toBe(true);
+    expect(set.has('c')).not.toBe(true);
+
+    resolveE();
+    await delay(1);
+
+    // `d` and `c` are the next nodes in the graph, so they should be resolved
+    // next.
+    expect(set.has('d')).toBe(true);
+    expect(set.has('c')).toBe(true);
+
+    resolveD();
+    await delay(1);
+
+    // `b` depends on `d` and `e`, so it should be resolved once they are.
+    expect(set.has('b')).toBe(true);
+    expect(set.has('a')).not.toBe(true);
+
+    resolveC();
+    resolveB();
+    await delay(1);
+
+    // `a` depends on `b` and `c`, so it should be resolved once they are.
+    expect(set.has('a')).toBe(true);
+
+    resolveA();
+    await parallelPromise;
   });
 });
