+++ packages/test-utils/src/index.ts
@@ -4,4 +4,5 @@ export * from './file-system.js';
 export * from './fixtures.js';
 export * from './json.js';
 export * from './mocks.js';
+export * from './worker.js';
 export * from './virtual-environment.js';
